Adding base for the block
