+++ package.json
@@ -10,7 +10,7 @@
   },
   "dependencies": {
     "@wordpress/api-fetch": "^2.2.8",
-    "@wordpress/block-library": "^2.2.16",
+    "@wordpress/block-library": "^2.14.1",
     "@wordpress/data": "^1.2.1",
     "@wordpress/html-entities": "^1.0.3",
     "@wordpress/i18n": "^1.2.3",
@@ -22,29 +22,29 @@
     "har-validator": "^5.1.3",
     "hex-to-rgba": "^1.0.2",
     "hoek": "^5.0.4",
-    "lazysizes": "^4.1.5",
-    "lodash": "^4.17.11",
+    "lazysizes": "^4.1.8",
+    "lodash": "^4.17.15",
     "md5": "^2.2.1",
     "moment": "^2.24.0",
-    "node-sass": "^4.10.0",
-    "querystringify": "^2.1.0",
+    "node-sass": "^4.13.1",
+    "querystringify": "^2.1.1",
     "randomatic": "^3.1.1",
-    "re-resizable": "^4.9.3",
+    "re-resizable": "^4.11.0",
     "react": "16.2.0",
     "react-lazy-load": "^3.0.13",
-    "react-lazyload": "^2.5.0",
+    "react-lazyload": "^2.6.5",
     "react-slick": "^0.23.2",
     "slick-carousel": "^1.8.1",
-    "sshpk": "^1.15.2",
+    "sshpk": "^1.16.1",
     "styled-components": "^3.4.10"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.5",
-    "@wordpress/babel-plugin-makepot": "^2.1.2",
+    "@babel/core": "^7.8.4",
+    "@wordpress/babel-plugin-makepot": "^2.1.3",
     "babel-cli": "^6.26.0",
     "babel-plugin-lodash": "^3.3.4",
     "babel-preset-env": "^1.7.0",
     "babel-preset-stage-2": "^6.24.1",
-    "grunt-wp-i18n": "^1.0.2"
+    "grunt-wp-i18n": "^1.0.3"
   }
 }
