+++ src/blocks/block-team/index.js
@@ -3,6 +3,7 @@
  */
 
 import './style.scss';
+import edit from './edit';
 
 /**
  * Internal block libraries
@@ -21,38 +22,7 @@ registerBlockType( 'lsx-blocks/lsx-team', {
 		__( 'team', 'lsx-blocks' ),
 		__( 'lsx', 'lsx-blocks' ),
 	],
-	edit: withSelect( select => {
-			return {
-				posts: select( 'core' ).getEntityRecords( 'postType', 'team', { per_page: 6 } )
-			};
-		} )( ( { posts, className, isSelected, setAttributes } ) => {
-			if ( ! posts ) {
-				return (
-					<p className={className} >
-						<Spinner />
-						{ __( 'Loading Posts', 'lsx-blocks' ) }
-
-					</p>
-				);
-			}
-			if ( 0 === posts.length ) {
-				return <p>{ __( 'No Posts', 'lsx-blocks' ) }</p>;
-			}
-			return (
-				<ul className={ className }>
-					{ posts.map( post => {
-						return (
-							<li>
-								<a className={ className } href={ post.link }>
-									{ post.title.rendered }
-								</a>
-							</li>
-						);
-					}) }
-				</ul>
-			);
-		} ) // end withAPIData
-	, // end edit
+	edit,
 	save() {
 		// Rendering in PHP
 		return null;
